src/main/java/com/sheen/joe/bankingsystem/controller/UserController.java
package com.sheen.joe.bankingsystem.controller;

import com.sheen.joe.bankingsystem.dto.UserResponseDto;
import com.sheen.joe.bankingsystem.service.UserService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.UUID;

@Slf4j
@RestController
@RequiredArgsConstructor
@RequestMapping("/api/v1/users")
public class UserController {

    private final UserService userService;

    @GetMapping(path = "/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<UserResponseDto> getById(@PathVariable("userId") UUID id) {
        UserResponseDto userResponseDto = userService.getUserById(id);
        log.info("User with ID: {} found", userResponseDto.id());
        return new ResponseEntity<>(userResponseDto, HttpStatus.OK);
    }

}
